api/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.30.1
azure-functions==1.18.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
pandas==2.1.1